\name{realBiomarker}
\docType{data}
\alias{realBiomarker}
\title{
A set of 54359 median gene expressions in log (base 2) scale 
}
\description{
This data set contains a set of 54359 log base 2 gene expression values from  a neonatal 
whole blood gene expression study described in Smith \emph{et al.} (2007). The data represent
the median of 28 microarrays corresponding to 28 control (healthy) patients of the neonatal study.
This data set is used as a base expressions set for simulating
biomarker data using \code{\link{simData}} function of the \code{\link{optBiomarker}} package.}

\usage{data(realBiomarker)}

\format{A vector of 54359 gene expressions in log (base 2) scale.}


\references{
Smith, C. L., Dickinson, P., Forster, T., Khondoker, M. R., Craigon, M., Ross, A., Storm, P., Burgess, S., Lacaze, P., Stenson, B. J.and  Ghazal, P. (2007). 
Quantitative assessment of whole blood RNA as a potential biomarker for infectious disease.
\emph{Analyst} \bold{132}, 1200--1209.   

Khondoker, M. R., Till T. Bachmann, T. T., Mewissen, M., Dickinson, P. \emph{et al.}(2010).
Multi-factorial analysis of class prediction error: estimating optimal number of biomarkers for various classification rules.
\emph{Journal of Bioinformatics and Computational Biology}, \bold{8}, 945-965.


}

\keyword{datasets}
